fix: Handle error when passed an undefined model
Fixes https://github.com/mongodb-js/connect-backbone-to-react/issues/8

I added a safety check when mounting or unmounting components for
undefined models. Since there's nothing we can do to a null model, we
should not attempt to add event listeners. If the model is actually
required, prop types should validate it, not a type error.