feat:Add props as second arg to mapModelsToProps
Implements https://github.com/mongodb-js/connect-backbone-to-react/issues/9

This is a nice feature for deriving which models to map into your state
based on existing props. It can be especially useful for retrieving a
model from a collection based on a value (as in the tests).

Just a two-line change in the main file for lots of increased
functionality.