Prototype using the React 16.3 Context API
In [React
16.3](https://reactjs.org/blog/2018/03/29/react-v-16-3.html#official-context-api)
they announced a new, official Context API. Here we can use this api to
simplify the `BackboneProvider` and the `connectBackboneToReact`
function. We no longer need to use soon to be deprecated lifecycle
methods and can instead just set up listeners for backbone events and
map models to props as previously.

Fortunately, we can accomplish this without changes to the
`connect-backbone-to-react` API, but unfortunately, it requires updating
to `react@^16.3.0` (along with `prop-types` and `react-dom`). This also
means that tests have to be updated to use the latest version of
`enzyme` which doesn't [yet
support](https://github.com/airbnb/enzyme/issues/1553) the new Context
API. Until then, we can leave this as a prototype.