fix: adds check for null model to componentWillReceiveProps
There are some cases where a Backbone model passed through CBR
might become undefined. (Consider the case where a view relies on
a secondary model as a "decorator", and the use of that "decorator"
model is toggled on or off.) Today, when a model becomes undefined,
CBR throws a TypeError in ComponentWillReceiveProps:

Uncaught TypeError: Cannot read property 'on' of undefined

This commit adds a check for null or undefined models before trying
to set event listeners in componentWillReceiveProps.