Checks this.context.models for models becoming undefined
Previously, we only checked this.props.models for models becoming
undefined in componentWillReceiveProps. This commit also checks
this.context.models, but only if a matching model was not found in
this.props.models -- because of the Object.assign in setModels,
if there are two models with the same mapKey, we prefer the one
in props.