fix: adds check for null model to componentWillReceiveProps (#21)
* fix: adds check for null model to componentWillReceiveProps

There are some cases where a Backbone model passed through CBR
might become undefined. (Consider the case where a view relies on
a secondary model as a "decorator", and the use of that "decorator"
model is toggled on or off.) Today, when a model becomes undefined,
CBR throws a TypeError in ComponentWillReceiveProps:

Uncaught TypeError: Cannot read property 'on' of undefined

This commit adds a check for null or undefined models before trying
to set event listeners in componentWillReceiveProps.

* Removes listeners from models which become undefined

* DRYs removeEventListener code

* Reference model directly in componentWillReceiveProps

* Checks this.context.models for models becoming undefined

Previously, we only checked this.props.models for models becoming
undefined in componentWillReceiveProps. This commit also checks
this.context.models, but only if a matching model was not found in
this.props.models -- because of the Object.assign in setModels,
if there are two models with the same mapKey, we prefer the one
in props.

* Adds test for model becoming undefined in componentWillReceiveProps

* Micro-optimize and clean componentWillReceiveProps

* Refactor undefined model test per review

* Adds testing for initial model listener bindings

* Adds tests: createListener and removeListener are called for decorator model

* Updates test titles and comments per review